requirements.txt
django>=6.0
djangorestframework>=3.17
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.9
python-decouple>=3.8
Pillow>=12.0
django-filter>=25.0
drf-spectacular>=0.29
anthropic>=0.100
gunicorn>=23.0
dj-database-url>=2.0
psycopg2-binary>=2.9
whitenoise>=6.7
cloudinary>=1.41
django-cloudinary-storage>=0.3
